requirements.txt
Flask
python-dotenv
pytesseract
Pillow
python-docx
PyMuPDF
google-generativeai
openpyxl
pandas